UniDjango/requirements.txt
-i https://pypi.org/simple
appscript==1.2.2 ; sys_platform == 'darwin'
asgiref==3.5.2 ; python_version >= '3.7'
babel==2.11.0 ; python_version >= '3.6'
certifi==2022.9.24 ; python_version >= '3.6'
charset-normalizer==2.1.1 ; python_full_version >= '3.6.0'
convertapi==1.5.0
django==4.1.3
docx2pdf==0.1.8
docxcompose==1.3.7
docxtpl==0.16.4
filestack-python==3.5.0
fpdf==1.7.2
idna==3.4 ; python_version >= '3.5'
jinja2==3.1.2 ; python_version >= '3.7'
lxml==4.9.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
markupsafe==2.1.1 ; python_version >= '3.7'
mega.py==1.0.8
pathlib==1.0.1
pip==22.3.1
pycryptodome==3.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
python-docx==0.8.11
pytz==2022.6
requests==2.28.1 ; python_version >= '3.7' and python_version < '4'
setuptools==65.6.3 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.3 ; python_version >= '3.5'
tenacity==5.1.5
tqdm==4.64.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
trafaret==2.0.2
urllib3==1.26.13 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
